fix/ - All DONE. ToDo - comments, code refactoring
