+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -8,8 +8,7 @@ import ru.practicum.shareit.booking.dto.BookingAddDto;
 import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.BookingMapper;
 import ru.practicum.shareit.booking.dto.State;
-import ru.practicum.shareit.item.dto.ItemMapper;
-import ru.practicum.shareit.item.dto.ItemOwnerDto;
+import ru.practicum.shareit.booking.model.Booking;
 
 import java.util.List;
