+++ src/main/java/ru/practicum/shareit/booking/BookingService.java
@@ -1,16 +1,20 @@
 package ru.practicum.shareit.booking;
 
+import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.booking.dto.BookingAddDto;
-import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.State;
+import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
+@Transactional(readOnly = true)
 public interface BookingService {
 
+    @Transactional
     Booking addBooking(BookingAddDto bookingDto, Long userId);
 
+    @Transactional
     Booking changeStatus(Long id, Boolean approved, Long userId);
 
     Booking findSpecificBooking(Long id, Long userId);
@@ -22,5 +26,4 @@ public interface BookingService {
     Booking findLastBooking(Item item);
 
     Booking findFutureBooking(Item item);
-//    Booking getPastBooking(Long itemId, Long bookerId);
 }
