+++ src/main/java/ru/practicum/shareit/item/dto/ItemMapper.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.item.dto;
 
-import ru.practicum.shareit.booking.Booking;
+import ru.practicum.shareit.booking.model.Booking;
 import ru.practicum.shareit.booking.BookingService;
 import ru.practicum.shareit.booking.dto.BookingMapper;
 import ru.practicum.shareit.item.ItemService;
