+++ src/main/java/ru/practicum/shareit/item/model/Comment.java
@@ -21,11 +21,11 @@ public class Comment {
     @NotBlank
     String text;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "item_id", nullable = false)
     private Item item;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "author_id", nullable = false)
     private User author;
